refactor: Point와 Product의 price를 정교한 값을 위해 타입을 BigDecimal로 변경

+++ src/main/java/cart/dao/ProductDao.java
@@ -27,7 +27,7 @@ public class ProductDao {
             new Product(
                     resultSet.getLong(ID),
                     resultSet.getString(NAME),
-                    resultSet.getInt(PRICE),
+                    resultSet.getBigDecimal(PRICE),
                     resultSet.getString(IMAGE_URL),
                     resultSet.getInt(STOCK)
             );
